the machine arm was OK,it can take multim meal

+++ STM32_Demo/utilities/include/Mini_hopper.h
@@ -1,8 +1,8 @@
 #ifndef __Mini_hopper_h
 #define __Mini_hopper_h
 
-extern uint16_t Coins_cnt; //一键退币计数
-extern uint16_t Coins_totoal; //硬币总数统计
+extern uint32_t Coins_cnt; //一键退币计数
+extern uint32_t Coins_totoal; //硬币总数统计
 extern uint8_t ErrorType;
 void InitMiniGPIO(void);
 u8 SendOutN_Coin(int num);
